[APPC-2231] Extraction of DB instantiations to another dagger module
